Moved loadAbiFromFile to helper

+++ utils/functionSignatures.js
@@ -4,21 +4,20 @@
  */
 
 require('dotenv').config()
-const Protocol = require('@graphprotocol/graph-cli/src/protocols/')
-const toolbox = require('gluegun/toolbox');
 const { Command } = require('commander');
 const program = new Command();
 const fs = require('fs-extra')
 program.version('0.0.1');
 const keccak256 = require('keccak256')
-const protocolInstance = new Protocol('ethereum')
-const ABI = protocolInstance.getABI()
+const { loadAbiFromFile } = require('./helpers')
 
 async function main() {
     const csvFile = './functionSignatures.csv'
-    fs.unlinkSync(csvFile)
+    if (fs.existsSync(csvFile)) {
+        fs.unlinkSync(csvFile)
+    }
     fs.createFile(csvFile)
-    fs.readdir('./abisForSignatures', async (err, files) => {
+    fs.readdir('./abis', async (err, files) => {
         files.forEach(async (file) => {
             console.log(file)
             const filePath = `abis/${file}`
@@ -35,20 +34,6 @@ async function main() {
             })
         });
     });
-}
-
-
-const loadAbiFromFile = async filename => {
-    let exists = await toolbox.filesystem.exists(filename)
-    if (!exists) {
-        throw Error(`File ${filename} does not exist.`)
-    } else if (exists === 'dir') {
-        throw Error('Path points to a directory, not a file.')
-    } else if (exists === 'other') {
-        throw Error('Not sure what this path points to.')
-    } else {
-        return await ABI.load('Contract', filename)
-    }
 }
 
 main()
